波紋攻撃に吹っ飛ばし追加
