+++ src/Assets/Prefabs/EnemyAct/RingNoizeWave.prefab
@@ -15,7 +15,7 @@ GameObject:
   - component: {fileID: -5342226985709423376}
   - component: {fileID: 854815500083653401}
   m_Layer: 10
-  m_Name: ShautWave
+  m_Name: RingNoizeWave
   m_TagString: Noize
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -104,7 +104,7 @@ BoxCollider:
     serializedVersion: 2
     m_Bits: 0
   m_LayerOverridePriority: 0
-  m_IsTrigger: 0
+  m_IsTrigger: 1
   m_ProvidesContacts: 0
   m_Enabled: 1
   serializedVersion: 3
@@ -151,3 +151,5 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   ScaleUpSpeed: 1
   ScaleUPMax: 10
+  AddPowerFlat: 1
+  AddPowerY: 1
