+++ src/Assets/Scenes/Test/EnemyObstructionTest.unity
@@ -1089,7 +1089,7 @@ Rigidbody:
   m_IsKinematic: 0
   m_Interpolate: 0
   m_Constraints: 112
-  m_CollisionDetection: 0
+  m_CollisionDetection: 2
 --- !u!136 &1014929416
 CapsuleCollider:
   m_ObjectHideFlags: 0
